requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.0
aiosignal==1.4.0
aiosqlite==0.22.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
discord==2.3.2
discord.py==2.6.4
dnspython==2.8.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.7.1
multidict==6.7.0
packaging==26.0
pillow==12.1.0
priority==2.0.0
propcache==0.4.1
psutil==7.2.1
Pygments==2.19.2
pymongo==4.16.0
python-dotenv==1.2.1
pytz==2023.3
Quart==0.20.0
requests==2.32.5
rich==14.3.1
setuptools==80.10.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wheel==0.46.3
wsproto==1.3.2
yarl==1.22.0
